Move date range picker to own folder
