+++ src/common/components/dateRangePicker/datePicker.scss
